requirements.txt
beautifulsoup4
python-arango
flask